# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Calorie Counter",
    "description": "Calorie Counter Node/Express API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerJWT": {
        "type": "http",
        "description": "JWT to use protected routes. Can be acquired using the API. You only need to paste the JWT. There is no need to write \"Bearer\" inside of the input box.",
        "scheme": "bearer",
        "in": "header",
        "name": "Authorization"
      }
    }
  },
  "security": [{
      "bearerJWT": []
  }],
  "host": "localhost:4000",
  "basePath": "/",
  "tags": [{
    "name": "Auth",
    "description": "Authenticated Services"
  }, {
    "name": "Calorie Logs",
    "description": "Logs for users to record their food intake"
  }, {
    "name": "Exercise Logs",
    "description": "Logs for users to record their exercise behavior"
  }, {
    "name": "Friends",
    "description": "Social media friend requests/logs"
  }, {
    "name": "Users",
    "description": "Interacting with user records"
  }],
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/v1/auth": {
      "post": {
        "tags": ["Auth"],
        "security": [],
        "summary": "Login to the API by receiving a valid JWT",
        "requestBody": {
          "required": true,
          "description": "User to be logged in",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/AuthPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "$ref": "#/definitions/Authenticated" }
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Email not found"
          }
        }
      }
    },
    "/api/v1/calories/createLog": {
      "post": {
        "tags": ["Calorie Logs"],
        "summary": "Create new Calorie Log",
        "requestBody": {
          "required": true,
          "description": "data relating to the calorie log being created",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/definitions/CalorieLogPayload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "array", "items": "objects" }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "un-authenticated"
          },
          "404": {
            "description": "Email not found"
          }
        }
      }
    },
    "/api/v1/calories/user/{userId}": {
      "get": {
        "tags": ["Calorie Logs"],
        "summary": "Get all Calorie Logs by a userId",
        "security": [],
        "parameters": [{
          "in": "path",
          "name": "userId",
          "required": true
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "array", "items": "objects" }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "un-authenticated"
          },
          "404": {
            "description": "Email not found"
          }
        }
      }
    },
    "/api/v1/calories/log/{logId}": {
      "get": {
        "tags": ["Calorie Logs"],
        "summary": "Get a single calorieLog by logId",
        "security": [],
        "parameters": [{
          "in": "path",
          "name": "logId",
          "required": true
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "un-authenticated"
          },
          "404": {
            "description": "Email not found"
          }
        }
      },
      "put": {
        "tags": ["Calorie Logs"],
        "summary": "Update Calorie Log",
        "parameters": [{
          "in": "path",
          "name": "logId",
          "required": true
        }],
        "requestBody": {
          "required": true,
          "description": "data relating to the calorie log being updated",
          "content": {
            "application/json": {
              "schema": {
                "properties": {
                  "food": { "type": "string" },
                  "calories": { "type": "integer" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "un-authenticated"
          },
          "404": {
            "description": "Email not found"
          }
        }
      },
      "delete": {
        "tags": ["Calorie Logs"],
        "summary": "Delete Calorie Log",
        "parameters": [{
          "in": "path",
          "name": "logId",
          "required": true
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "object" }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "un-authenticated"
          },
          "404": {
            "description": "Email not found"
          }
        }
      }
    },
    "/api/v1/calories/daterange/{begin}/{end}": {
      "get": {
        "tags": ["Calorie Logs"],
        "summary": "Get all calorie logs in a specific time frame",
        "security": [],
        "parameters": [{
          "in": "path",
          "name": "begin",
          "required": true
        }, {
          "in": "path",
          "name": "end",
          "required": true
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "array", "items": "objects" }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "un-authenticated"
          },
          "404": {
            "description": "Email not found"
          }
        }
      }
    },
    "/api/v1/calories/food/{foodName}": {
      "get": {
        "tags": ["Calorie Logs"],
        "summary": "Get all calorie logs with a specific food name",
        "security": [],
        "parameters": [{
          "in": "path",
          "name": "foodName",
          "required": true
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "array", "items": "objects" }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "un-authenticated"
          },
          "404": {
            "description": "Email not found"
          }
        }
      }
    },
    "/api/v1/calories/calorierange/{begin}/{end}": {
      "get": {
        "tags": ["Calorie Logs"],
        "summary": "Get all calorie logs in a specific calorie range",
        "security": [],
        "parameters": [{
          "in": "path",
          "name": "begin",
          "required": true
        }, {
          "in": "path",
          "name": "end",
          "required": true
        }],
        "responses": {
          "200": {
            "description": "OK",
            "schema": { "type": "array", "items": "objects" }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "un-authenticated"
          },
          "404": {
            "description": "Email not found"
          }
        }
      }
    }
  },
  "definitions": {
    "Authenticated": {
      "required": [
        "user",
        "jwt"
      ],
      "properties": {
        "user": { "type": "object" },
        "jwt": { "type": "string" }
      }
    },
    "AuthPayload": {
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": { "type": "string" },
        "password": { "type": "string" }
      }
    },
    "CalorieLogPayload": {
      "required": [
        "userId",
        "food",
        "calories",
        "timeOfDay"
      ],
      "properties": {
        "userId": { "type": "string" },
        "food": { "type": "string" },
        "calories": { "type": "integer" },
        "timeOfDay": { "type": "string" }
      }
    }
  }
}